+++ Gruntfile.js
@@ -34,9 +34,7 @@ module.exports = function (grunt) {
             all: {
                 src: [ // list all sources in dependency order
                     'src/MVC-Core.st', 
-                    'src/App.st', 
-                    'src/MVC-Iterables.st', 
-                    'src/MVC-Routeables.st'
+                    'src/App.st'
                 ],
                 amd_namespace: 'amber-mvc',
                 libraries: [ 'SUnit', 'Web',